Feature/sync_progress (#40)
It adding the sync-progress feature which is announcing its sync progress in the _sync-method in the key verificationprogress of the getblockchaininfo call.
In order to not have circular dependencies which caused trouble installing with pip3 install -e ".[test]" , i removed all occurrences of specter-desktop.
Because of that, the tests, which were broken anyway, needed to be skipped and/or not executed in other ways which needed specter-desktop stuff.
The most essential things have been copy and pasted in the util_specter.py but i was not shy to exspcially skip the wallet_rescan test which never worked in the CI anyway and is now only there to inspire new tests.
Instead i tried to introduce bdk and get the integration tests going which are using Nigiri. So now we have at least a working Nigiri integration test setup (see basic.py) which is ready to have more tests.